+++ Hooks/PacketEvents/src/main/java/fr/maxlego08/menu/hooks/packetevents/listener/PacketAnimationListener.java
@@ -62,6 +62,7 @@ public class PacketAnimationListener implements PacketListener {
             WrapperPlayServerOpenWindow wrapper = new WrapperPlayServerOpenWindow(event);
             int containerId = wrapper.getContainerId();
             Player player = event.getPlayer();
+            if (player == null) return;
             UUID playerUniqueId = player.getUniqueId();
 
             PlayerAnimationData data = this.playerAnimationData.get(playerUniqueId);
@@ -97,6 +98,7 @@ public class PacketAnimationListener implements PacketListener {
             }
         } else if (packetType == PacketType.Play.Server.CLOSE_WINDOW){
             Player player = event.getPlayer();
+            if (player == null) return;
             Inventory topInventory = CompatibilityUtil.getTopInventory(player);
             if (topInventory == null) {
                 return;
@@ -118,6 +120,7 @@ public class PacketAnimationListener implements PacketListener {
         } else if (packetType == PacketType.Play.Server.WINDOW_ITEMS){
             WrapperPlayServerWindowItems wrapper = new WrapperPlayServerWindowItems(event);
             Player player = event.getPlayer();
+            if (player == null) return;
             UUID playerUniqueId = player.getUniqueId();
 
             int windowId = wrapper.getWindowId();